refactored out separate execute_async func (relates to #2)

+++ tests/jktebop/test_jktebop.py
@@ -35,17 +35,14 @@ ${out_filename}
     #
     def test_execute_task_valid(self):
         """ Test execute_task(all params for full task 2 happy path) check all processing """
-        in_file, out_file = self._test_specific_in_and_out_filenames()
+        # Clear out any artefacts from previous run of this test & generate a new valid in file
+        in_file, out_file = self._test_specific_in_and_out_filenames(unlink_existing=True)
+        self._write_task2_in_file(in_file, out_file=out_file)
         cleanup_pattern = f"{in_file.stem}.*"
         stdout_to = StringIO()
 
-        # Clear out any artefacts from previous run of this test & generate a new valid in file
-        for file in [in_file, out_file]:
-            file.unlink(missing_ok=True)
-        self._write_task2_in_file(in_file, out_file=out_file)
-
         # Run the task and capture the response in a list[str]
-        out_lines = [*jktebop.execute_task(in_file, out_file, cleanup_pattern, True, stdout_to)]
+        out_lines = list(jktebop.execute_task(in_file, out_file, cleanup_pattern, True, stdout_to))
 
         # Assert processing
         model = np.loadtxt(out_lines, comments="#", dtype=np.double, unpack=True)
@@ -57,41 +54,81 @@ ${out_filename}
 
     def test_execute_task_valid_no_cleanup(self):
         """ Test execute_task(happy path, but no cleanup_pattern) check in/out files left behind """
-        in_file, out_file = self._test_specific_in_and_out_filenames()
-
         # Clear out any artefacts from previous run of this test & generate a new valid in file
-        for file in [in_file, out_file]:
-            file.unlink(missing_ok=True)
+        in_file, out_file = self._test_specific_in_and_out_filenames(unlink_existing=True)
         self._write_task2_in_file(in_file, out_file=out_file)
 
-        # Run the task and capture the response in a list[str]
+        # Run the task and wait for completion (based on the generator terminating)
         list(jktebop.execute_task(in_file, out_file, None))
 
         # Assert in and out files not cleaned down
         self.assertTrue(in_file.exists())
         self.assertTrue(out_file.exists())
 
+    def test_execute_task_valid_test_stdout_to(self):
+        """ Test execute_task(happy path specifically testing stdout_to) """
+        # Clear out any artefacts from previous run of this test & generate a new valid in file
+        in_file, out_file = self._test_specific_in_and_out_filenames(unlink_existing=True)
+        self._write_task2_in_file(in_file, out_file=out_file)
+        cleanup_pattern = f"{in_file.stem}.*"
+        stdout_to = StringIO()
+
+        # Run the task and wait for completion (based on the generator terminating)
+        list(jktebop.execute_task(in_file, out_file, cleanup_pattern, stdout_to=stdout_to))
+
+        # Look for text indicating that the processing has worked through successfully
+        console_text = stdout_to.getvalue()
+        self.assertIn(f"Opened new lightcurve file:  {out_file.name}", console_text)
+
     def test_execute_task_invalid_inc_value(self):
         """ Test execute_task(in_file has invalid inc value) so raises CalledProcessError """
-        in_file, out_file = self._test_specific_in_and_out_filenames()
-        cleanup_pattern = f"{in_file.stem}.*"
-
         # Clear artefacts from previous run of this test & generate a new in file with invalid inc
-        for file in [in_file, out_file]:
-            file.unlink(missing_ok=True)
+        in_file, out_file = self._test_specific_in_and_out_filenames(unlink_existing=True)
         self._write_task2_in_file(in_file, inc=45.0, out_file=out_file) # range for inc (50, 140)
+        cleanup_pattern = f"{in_file.stem}.*"
 
         # Assert we get an error and the output contains text indicating what the problem may be
         with self.assertRaises(CalledProcessError) as ect:
             list(jktebop.execute_task(in_file, out_file, cleanup_pattern))
         self.assertIn("### ERROR: the value of INCLNATION", ect.exception.output)
 
+    def test_execute_task_missing_in_file(self):
+        """ Test execute_task(in_file has invalid inc value) so raises CalledProcessError """
+        # Clear artefacts with these names but do not generate anything
+        in_file, out_file = self._test_specific_in_and_out_filenames(unlink_existing=True)
+        with self.assertRaises(FileNotFoundError):
+            list(jktebop.execute_task(in_file, out_file, None))
+
+
+    #
+    # Tests execute_task_async(in_filename) -> Popen[str]:
+    #
+    def test_execute_task_async_valid_in_file(self):
+        """ Test execute_task_async(in_file is valid) returns valid Popen which can be waited on """
+        in_file, out_file = self._test_specific_in_and_out_filenames(unlink_existing=True)
+        self._write_task2_in_file(in_file, out_file=out_file)
+
+        with jktebop.execute_task_async(in_file) as proc:
+            self.assertIsNone(proc.poll())  # Will return None if still running
+            proc.wait()
+
+    def test_execute_task_async_missing_in_file(self):
+        """ Test execute_task_async(in_file is mission) so raises FileNotFoundError """
+        # Clear artefacts with these names but do not generate anything
+        in_file, _ = self._test_specific_in_and_out_filenames(unlink_existing=True)
+        with self.assertRaises(FileNotFoundError):
+            jktebop.execute_task_async(in_file)
+
+
 
     # Helper functions
-    def _test_specific_in_and_out_filenames(self):
+    def _test_specific_in_and_out_filenames(self, unlink_existing: bool=False):
         """ Generates matching in and out file names based on the calling function name """
         in_file = self._jktebop_dir / f"{inspect.stack()[1].function}.in"
         out_file = self._jktebop_dir / f"{in_file.stem}.out"
+        if unlink_existing:
+            for file in [in_file, out_file]:
+                file.unlink(missing_ok=True)
         return (in_file, out_file)
 
     def _write_task2_in_file(self, filename: Path, out_file: Path, inc: float=90.0):
